refactor: [#472] move generic HTTP client
to the generic HTTP mod.
